add midleware adn fix create order product

+++ routes/api.php
@@ -21,13 +21,13 @@ Route::get('/product/details', [ProductController::class, 'productdetails']);
 Route::get('/list-product-type',[ProductController::class, 'getListProductType']);
 Route::get('/list-variant-type', [ProductController::class, 'getListVariantType']);
 
-Route::post('/upload-image', [ProductController::class, 'uploadImage']);
-Route::post('/insert-product', [ProductController::class, 'insertProduct']);
-Route::post('/insert-product-type', [ProductController::class, 'createProductType']);
-Route::post('/insert-variant-type', [ProductController::class, 'createVariantType']);
+Route::post('/upload-image', [ProductController::class, 'uploadImage'])->middleware('auth.admin');
+Route::post('/insert-product', [ProductController::class, 'insertProduct'])->middleware('auth.admin');
+Route::post('/insert-product-type', [ProductController::class, 'createProductType'])->middleware('auth.admin');
+Route::post('/insert-variant-type', [ProductController::class, 'createVariantType'])->middleware('auth.admin');
 
-Route::put('/update-product/{id}', [ProductController::class, 'updateProduct']);
-Route::delete('/deleted-product/{id}', [ProductController::class, 'deleteProduct']);
+Route::put('/update-product/{id}', [ProductController::class, 'updateProduct'])->middleware('auth.admin');
+Route::delete('/deleted-product/{id}', [ProductController::class, 'deleteProduct'])->middleware('auth.admin');
 
 Route::post('/admin/register', [AuthController::class, 'register']);
 Route::post('/admin/login', [AuthController::class, 'login']);
@@ -35,7 +35,8 @@ Route::post('/admin/logout', [AuthController::class, 'logout'])->middleware('aut
 
 Route::get('/list-order', [BookingController::class, 'getListOrder']);
 Route::get('/detail-order',[BookingController::class, 'getDetilsOrder']);
-Route::post('/update-status-order',[BookingController::class, 'updateStatusOrder']);
+Route::post('/update-status-order',[BookingController::class, 'updateStatusOrder'])->middleware('auth.admin');
+Route::post('/create-orders', [BookingController::class, 'createOrder'])->middleware('auth.admin');
 
 Route::middleware('auth:sanctum')->group(function () {
     Route::get('/admin/dashboard', [AdminController::class, 'dashboard']);
